refactor: move logging to app/utils/logging.py
